рабочий прокси опрос rs заготовка dio

+++ server/src/tests.devices/result/send.sms.ts
@@ -1,7 +1,10 @@
 import { Sms, Device } from '../../schema';
 
-const cmd = require('node-cmd')
-
+import cmd  from 'node-cmd'
+import { isUndefined } from 'util';
+import { RTUproxyReguest } from '../modbus.test';
+import { TCPproxyReguest } from '../modbusProxy/TCP.proxy';
+//echo "-w=26:0 0 0 1 0 1 0" >/sys/devices/virtual/misc/mtgpio/pin
 const serialportgsm = require('serialport-gsm')
 
 const modem = serialportgsm.Modem()
@@ -22,24 +25,37 @@ const options = {
     customInitCommand: '',
     logger: console
 }
-cmd.get("ps|grep ril", (err,data,stderr)=>{
+const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
+
+const init = ()=> cmd.get("ps|grep rild", async(err,data,stderr)=>{
+     
+
     if (!err) {          
-      const pid = String(data).substr(10,6)
+      const pid = data.substr(10,6)
       console.log("pid:",pid)
-    //  if(pid){
-              cmd.get("kill -STOP "+ pid, async (err,data,stderr)=>{
-                const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
+      if( !(parseInt( pid ) > 0) ){
+        
+        await sleep(1000)
+        setImmediate(() => {
+            init()
+        })
+        return
+      }
+      await sleep(10000)        
+      cmd.get("kill -STOP "+ pid, async (err,data,stderr)=>{
                 await sleep(3000)
                 if (!err) {
                     modem.open('/dev/radio/atci1', options,(err,res)=>{
                         if(err){
                             console.error('reboot',err)
-                            cmd.get('reboot')
+                            cmd.run('reboot')
                         }
                         else
                         cmd.get('svc wifi disable && service call wifi 29 i32 0 i32 1'/* && stop ril-daemon*/,(err, data, stderr)=>{
                             if (!err) {
                                console.log(data)
+
+                               setTimeout(()=>cmd.run('stop zygote'),50000)
                             } else {
                                console.log('error', err)
                             }
@@ -58,11 +74,12 @@ cmd.get("ps|grep ril", (err,data,stderr)=>{
 })
 
 
+init()
 
   modem.on('open', () => {
 
 
-            getNetworkSignal()
+           setInterval(getNetworkSignal,40000)
             modem.setModemMode(  (msg,err)=>{
                 if(err)
                     console.error(err)
@@ -72,23 +89,43 @@ cmd.get("ps|grep ril", (err,data,stderr)=>{
   
 })
    
-  
+const lamp = { intervalId:undefined, state:false }
+process.on('exit',(code)=>cmd.run('echo "-w=26:0 0 0 0 0 1 0" > /sys/devices/virtual/misc/mtgpio/pin')) // погасить
 const getNetworkSignal = ()=>{
-    //console.log(getNetworkSignal)
+    if(!(RTUproxyReguest.length||TCPproxyReguest.length))
     modem.getNetworkSignal(async(result, error)=>{
-
+       
         if(!error){
-            console.dir(result)
+            const q=parseInt(result.data.signalQuality)
+           // console.dir( result )
+            //console.log( 'q:',q,lamp.intervalId )
+           
+            if( q > 6 && q <= 30 ){     
+                clearInterval( lamp.intervalId ) 
+                lamp.intervalId = undefined
+                cmd.run('echo "-w=26:0 0 0 1 0 1 0" > /sys/devices/virtual/misc/mtgpio/pin') // зажеч   
+               }else{
+              if( q <= 6  && isUndefined( lamp.intervalId ))
+                lamp.intervalId = setInterval(()=>{
+                    lamp.state =  !lamp.state
+                    if( lamp.state ) 
+                    cmd.run('echo "-w=26:0 0 0 1 0 1 0" > /sys/devices/virtual/misc/mtgpio/pin') // зажеч   
+                    else
+                    cmd.run('echo "-w=26:0 0 0 0 0 1 0" > /sys/devices/virtual/misc/mtgpio/pin') // погасить
+                    
+                    },200) as any
+               else{
+                clearInterval(lamp.intervalId)
+                cmd.run('echo "-w=26:0 0 0 0 0 1 0" > /sys/devices/virtual/misc/mtgpio/pin') // погасить
+               }
+            }
+
         }else
             console.log('getNetworkSignal:',error)
-
-        const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
         
-        await sleep(10000)
 
-        setImmediate(() => {
-            getNetworkSignal()
-        })
+
+
     })
 }
 
@@ -97,7 +134,7 @@ export function sendSMS(sms:Sms,device?:Device){
     const interval = setInterval((sms:Sms,device:Device)=>{
          const text=device.name +':'+sms.text
         for(const mumber of sms.numbers)
-         if(mumber)modem.sendSMS( mumber, text, false, 
+         if( mumber )modem.sendSMS( mumber, text, false, 
             (result)=>{
                 function* x(){
                     if(result.status==='success')
